Register native auditlogformatter in TinyGo
